+++ Assets/_Dev/Game/Scripts/Entities/Turrets/Base/Turret.cs
@@ -4,7 +4,7 @@ namespace _Dev.Game.Scripts.Entities.Turrets.Base
 {
     public abstract class Turret : IAttacker
     {
-        public abstract int GetAttackInterval { get; }
+        public abstract float GetAttackInterval { get; }
         public abstract int GetAttackDamage { get; }
         public abstract void Attack();
     }
